+++ spec/integration/braintree/dispute_gateway_spec.js
@@ -1,7 +1,6 @@
- 'usestrict';
+'use strict';
 
 let fs = require('fs');
-let sinon = require('sinon');
 let CreditCardNumbers = require('../../../lib/braintree/test/credit_card_numbers').CreditCardNumbers;
 let Dispute = require('../../../lib/braintree/dispute').Dispute;
 let DocumentUpload = require('../../../lib/braintree/document_upload').DocumentUpload;