+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Remove deprecated iDEAL support
 * Remove deprecated `connect` method
 * Config class now throws an error when token environment does not match passed environment
+* Remove `GrantedPaymentInstrumentUpdate` (deprecated in 2.16.0)
 
 ## Unreleased
 * Add `networkTransactionId` to `CreditCardVerification`