+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## WIP: next major version
 * Add `RequestTimeoutError` and `GatewayTimeoutError` exceptions
 Breaking Changes:
-* Remove deprecated iDEAL, coinbase support
+* Remove deprecated iDEAL, Coinbase, and Transparent Redirect
 * Remove deprecated `connect` method
 * Config class now throws an error when token environment does not match passed environment
 * Remove `GrantedPaymentInstrumentUpdate` (deprecated in 2.16.0)